# openapi/yodlee-ops.yaml
openapi: 3.0.3
info:
  title: Yodlee Ops API
  version: 1.0.0
  description: API for Yodlee Ops service

servers:
  - url: /yodlee-ops/api/v1

paths:
  # note: most code generators don't work with server-sent events, this is documentation only.
  /events/taillog:
    get:
      security:
        - bearerAuth: [ ]
      summary: Stream tail logs.
      description: Stream logs for real-time ingestion events via Server-Sent Events.
      operationId: streamLogs
      responses:
        '200':
          description: Successful SSE stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/FiObject"

        '500':
          description: Server error

  /fimetadata/prefix:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get FI metadata
      description: Returns a list of FI metadata records filtered by query parameters.
      operationId: getFiMetadataByPrefix
      parameters:
        - name: prefix
          in: query
          description: Prefix to perform search with.
          required: true
          schema:
            type: string
          example: "p1/Y/243512"
        - name: cursor
          in: query
          description: Cursor to paginate by. Remember to url encode it, ListObjectsV2 cursors contain unsafe URL characters.
          required: false
          schema:
            type: string
          example: "cDEvMS8xMDAvMjAyNS0wNi0xMw=="
        - name: subject
          in: query
          description: Subject filter value.
          required: true
          schema:
            $ref: "#/components/schemas/FiSubject"
          example: "accounts"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListFiMetadataResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResp"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResp"

  /fimetadata/profiles:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get FI metadata
      description: Returns a list of FI metadata records filtered by query parameters.
      operationId: getFiMetadataByProfiles
      parameters:
        - name: profileIDs
          in: query
          description: Profile IDs to filter by (comma-separated string).
          required: true
          schema:
            type: string
          example: "p1,p2"
        - name: cursor
          in: query
          description: Cursors to paginate by (comma-seperated string). Remember to url encode it, ListObjectsV2 cursors contain unsafe URL characters.
          required: true
          schema:
            type: string
          example: "cDEvMS8xMDAvMjAyNS0wNi0xMw==,cDEvMS8xMDAvMjAyNS0wNi0xMw=="
        - name: subject
          in: query
          description: Subject filter value.
          required: true
          schema:
            $ref: "#/components/schemas/FiSubject"
          example: "accounts"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListFiMetadataResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResp"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResp"

  /fiobject:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get FI object
      description: Returns the FI object data for the given key and subject.
      operationId: getFiObject
      parameters:
        - name: key
          in: query
          description: Key identifying the FI object.
          required: true
          schema:
            type: string
          example: "p1/1/10/100/2025-06-12"
        - name: subject
          in: query
          description: Subject filter value.
          required: true
          schema:
            $ref: "#/components/schemas/FiSubject"
          example: "accounts"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FiObject"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResp"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResp"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    FiObject:
      required:
        - profileId
        - timestamp
        - originTopic
        - data
      type: object
      properties:
        profileId:
          type: string
          example: "p1"
        timestamp:
          type: string
          format: date-time
          example: "2026-02-19T04:33:11Z"
        originTopic:
          type: string
          example: "acctResponseTopic"
        data:
          type: "object"
          description: Arbitrary JSON payload
          additionalProperties: true
          example:
            accountId: "100"
            balance: 5000

    FiSubject:
      type: string
      description: Subject filter value
      enum:
        - connections
        - accounts
        - transactions
        - holdings

    OpsFiMetadata:
      type: object
      required:
        - subject
        - key
        - lastModified
        - profileID
        - providerAccountID
        - partyIDTypeCd
        - accountID
        - elementID
        - lastUpdated
      properties:
        subject:
          $ref: "#/components/schemas/FiSubject"
        key:
          type: string
          example: "p1/1/10/100/2025-06-12"
        lastModified:
          type: string
          format: date-time
          example: "2026-02-19T04:33:11Z"
        profileID:
          type: string
          example: "p1"
        providerAccountID:
          type: string
          example: "1"
        partyIDTypeCd:
          type: string
          example: "10"
        accountID:
          type: string
          example: "100"
        elementID:
          type: string
          example: "1000"
        lastUpdated:
          type: string
          format: date-time
          example: "2025-06-12T00:00:00Z"

    ListFiMetadataResponse:
      type: object
      required:
        - opsFiMetadata
        - cursor
      properties:
        opsFiMetadata:
          type: array
          items:
            $ref: "#/components/schemas/OpsFiMetadata"
        cursor:
          type: string
          description: Comma seperated cursor that can be passed back into the list fi metadata call.
          additionalProperties:
            type: string
          example: "cDEvMS8xMDAvMjAyNS0wNi0xMw==,cDEvMS8xMDAvMjAyNS0wNi0xMw=="

    ErrorCode:
      type: string
      description: Machine readable error code
      enum:
        - PROFILE_ID_CURSOR_LENGTH
        - KEY_NOT_FOUND
        - FATAL_ERROR

    ErrorResp:
      type: object
      required:
        - errorCode
        - errorDesc
      properties:
        errorCode:
          $ref: "#/components/schemas/ErrorCode"
        errorDesc:
          type: string
          example: "Invalid query parameter"