# services/api.go
package svc

import (
	"context"
	"errors"
	"fmt"
	"io"
	"log/slog"
	"net/http"
	"strings"
	"time"

	"yodleeops/infra"
	openapi "yodleeops/openapi/sources"

	"github.com/go-faster/jx"
	"github.com/google/uuid"
	httpSwagger "github.com/swaggo/http-swagger/v2"
)

const ApiUrl = "/yodlee-ops/api/v1"

func RootMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		trace := r.Header.Get("X-trace")
		if trace == "" {
			trace = uuid.NewString()
		}
		r = r.WithContext(context.WithValue(r.Context(), "trace", trace))

		next.ServeHTTP(w, r)
	})
}

func MakeServeMux(app *State) *http.ServeMux {
	mux := http.NewServeMux()

	oGenServer, err := openapi.NewServer(&FiOpsAPIHandler{State: app}, &FiOpsSecurityHandler{State: app})
	if err != nil {
		panic(fmt.Sprintf("failed to make root: %v", err))
	}

	// ogen generated api: a trailing slash is needed for prefix matching. strip prefix allows ogen to do exact path matching.
	mux.Handle(ApiUrl+"/", http.StripPrefix(ApiUrl, RootMiddleware(oGenServer)))

	// streaming api: cannot be generated by ogen.
	mux.Handle(fmt.Sprintf("GET %s/events/taillog", ApiUrl), RootMiddleware(http.HandlerFunc(app.StreamFiObjectLogs)))

	// dedicated swagger-ui page: served from the openapi static directory.
	mux.HandleFunc("GET /swagger/yaml/yodlee-ops.yaml", func(w http.ResponseWriter, r *http.Request) {
		http.ServeFile(w, r, "./openapi/yodlee-ops.yaml")
	})
	mux.Handle("GET /swagger/", httpSwagger.Handler(
		httpSwagger.URL("/swagger/yaml/yodlee-ops.yaml"),
	))

	// healthcheck.
	mux.HandleFunc("GET /ping", func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(http.StatusOK)
	})

	return mux
}

func (state *State) AuthorizeToken(ctx context.Context, token string) (context.Context, error) {
	// todo: implement token validation here. in this implementation, Authorization allows all users.
	return ctx, nil
}

type FiOpsSecurityHandler struct {
	*State
}

func (h *FiOpsSecurityHandler) HandleBearerAuth(ctx context.Context, _ openapi.OperationName, t openapi.BearerAuth) (context.Context, error) {
	return h.AuthorizeToken(ctx, t.Token)
}

func FFormatEvent(w io.Writer, topic string, msg string) {
	fmt.Fprintf(w, "event: %s\ndata: %s\n\n", topic, msg)
}

func (state *State) StreamFiObjectLogs(w http.ResponseWriter, r *http.Request) {
	authorization := r.Header.Get("Authorization")
	ctx, err := state.AuthorizeToken(r.Context(), authorization)
	if err != nil {
		http.Error(w, "access denied", http.StatusUnauthorized)
		return
	}

	values := r.URL.Query()
	profileIDs := strings.Split(values.Get("prefix"), ",")
	subjects := strings.Split(values.Get("subjects"), ",")

	var topics []infra.Topic
	for _, subject := range subjects {
		switch openapi.FiSubject(subject) {
		case openapi.FiSubjectConnections:
			topics = append(topics, infra.CnctRefreshTopic, infra.CnctResponseTopic)
		case openapi.FiSubjectAccounts:
			topics = append(topics, infra.AcctRefreshTopic, infra.AcctResponseTopic)
		case openapi.FiSubjectTransactions:
			topics = append(topics, infra.TxnRefreshTopic, infra.TxnResponseTopic)
		case openapi.FiSubjectHoldings:
			topics = append(topics, infra.HoldRefreshTopic, infra.HoldResponseTopic)
		}
	}

	w.Header().Set("Content-Type", "text/event-stream")
	w.Header().Set("Cache-Control", "no-cache")
	w.Header().Set("Connection", "keep-alive")

	slog.InfoContext(ctx, "begin log tail streaming", "subjects", subjects, "topics", topics, "prefix", profileIDs)

	f, ok := w.(http.Flusher)
	if !ok {
		http.Error(w, "streaming unsupported", http.StatusMethodNotAllowed)
		return
	}

	FFormatEvent(w, "meta", "init")
	f.Flush()

	subChan := state.FiMessageBroadcaster.Subscribe(SubscriberFilter{Topics: topics, ProfileIDs: profileIDs})

	go func() {
		defer state.FiMessageBroadcaster.Unsubscribe(subChan)
		<-ctx.Done()
	}()

	ticker := time.NewTicker(15 * time.Second)
	for {
		select {
		case msg, ok := <-subChan:
			if !ok {
				return
			}
			FFormatEvent(w, "log", msg)
		case <-ticker.C:
			FFormatEvent(w, "meta", "ping")
		}
		f.Flush()
	}
}

func BucketFromSubject(subject openapi.FiSubject) infra.Bucket {
	var bucket infra.Bucket
	switch subject {
	case openapi.FiSubjectConnections:
		bucket = infra.CnctBucket
	case openapi.FiSubjectAccounts:
		bucket = infra.AcctBucket
	case openapi.FiSubjectTransactions:
		bucket = infra.TxnBucket
	case openapi.FiSubjectHoldings:
		bucket = infra.HoldBucket
	default:
		bucket = infra.CnctBucket
	}
	return bucket
}

func mapOpsFiMetadata(opsFiMetadata []OpsFiMetadata, subject openapi.FiSubject) []openapi.OpsFiMetadata {
	opsFiMetadataResp := make([]openapi.OpsFiMetadata, 0)
	for _, metadata := range opsFiMetadata {
		opsFiMetadataResp = append(opsFiMetadataResp, openapi.OpsFiMetadata{
			Subject:           subject,
			Key:               metadata.Key,
			LastModified:      metadata.LastModified,
			ProfileID:         metadata.ProfileID,
			ProviderAccountID: metadata.ProviderAccountID,
			PartyIDTypeCd:     metadata.PartyIDTypeCd,
			AccountID:         metadata.AccountID,
			ElementID:         metadata.ElementID,
			LastUpdated:       metadata.LastUpdated,
		})
	}
	return opsFiMetadataResp
}

type FiOpsAPIHandler struct {
	*State
}

var _ openapi.Handler = (*FiOpsAPIHandler)(nil)

const InternalServerErrorMessage = "internal server error"

func (h *FiOpsAPIHandler) GetFiMetadataByPrefix(ctx context.Context, params openapi.GetFiMetadataByPrefixParams) (openapi.GetFiMetadataByPrefixRes, error) {
	appCtx := Context{Context: ctx, State: h.State}

	internalServerError := func(err error) *openapi.GetFiMetadataByPrefixInternalServerError {
		slog.ErrorContext(ctx, "failed to list fi metadata by prefix", "err", err, "params", params)
		return &openapi.GetFiMetadataByPrefixInternalServerError{ErrorCode: openapi.ErrorCodeFATALERROR, ErrorDesc: InternalServerErrorMessage}
	}

	bucket := BucketFromSubject(params.Subject)

	opsFiMetadata, nextCursor, err := ListFiMetadataByPrefix(appCtx, bucket, params.Prefix, params.Cursor.Value)
	if err != nil {
		return internalServerError(err), nil
	}

	return &openapi.ListFiMetadataResponse{
		OpsFiMetadata: mapOpsFiMetadata(opsFiMetadata, params.Subject),
		Cursor:        nextCursor,
	}, nil
}

func (h *FiOpsAPIHandler) GetFiMetadataByProfiles(ctx context.Context, params openapi.GetFiMetadataByProfilesParams) (openapi.GetFiMetadataByProfilesRes, error) {
	appCtx := Context{Context: ctx, State: h.State}

	internalServerError := func(err error) *openapi.GetFiMetadataByProfilesInternalServerError {
		slog.ErrorContext(ctx, "failed to list fi metadata by profiles", "err", err, "params", params)
		return &openapi.GetFiMetadataByProfilesInternalServerError{ErrorCode: openapi.ErrorCodeFATALERROR, ErrorDesc: InternalServerErrorMessage}
	}

	profileIDs := strings.Split(params.ProfileIDs, ",")
	bucket := BucketFromSubject(params.Subject)

	var arrayCursor []string
	if params.Cursor == "" {
		// default: the first cursor is an empty ContinuationToken for each profileID
		arrayCursor = MakeFirstCursor(len(profileIDs))
	} else {
		arrayCursor = strings.Split(params.Cursor, ",")
	}

	if len(profileIDs) != len(arrayCursor) {
		return &openapi.GetFiMetadataByProfilesBadRequest{
			ErrorCode: openapi.ErrorCodePROFILEIDCURSORLENGTH,
			ErrorDesc: "prefix and cursor must be the same length",
		}, nil
	}
	var pairs []ListFiMetadataQuery
	for i, profileID := range profileIDs {
		pairs = append(pairs, ListFiMetadataQuery{ProfileID: profileID, ContinuationToken: arrayCursor[i]})
	}

	fiMetadataResults, err := ListFiMetadataByProfileIDs(appCtx, bucket, pairs)
	if err != nil {
		return internalServerError(err), nil
	}

	return &openapi.ListFiMetadataResponse{
		OpsFiMetadata: mapOpsFiMetadata(fiMetadataResults.OpsFiMetadata, params.Subject),
		Cursor:        fiMetadataResults.Cursor,
	}, nil
}

func (h *FiOpsAPIHandler) GetFiObject(ctx context.Context, params openapi.GetFiObjectParams) (openapi.GetFiObjectRes, error) {
	appCtx := Context{Context: ctx, State: h.State}

	internalServerError := func(err error) *openapi.GetFiObjectInternalServerError {
		slog.ErrorContext(ctx, "failed to get fi object", "err", err)
		return &openapi.GetFiObjectInternalServerError{ErrorCode: openapi.ErrorCodeFATALERROR, ErrorDesc: InternalServerErrorMessage}
	}

	keyInput := params.Key
	bucketInput := BucketFromSubject(params.Subject)

	opsFiObject, err := GetFiObject(appCtx, bucketInput, keyInput)
	if errors.Is(err, ErrKeyNotFound) {
		return &openapi.GetFiObjectNotFound{
			ErrorCode: openapi.ErrorCodeKEYNOTFOUND,
			ErrorDesc: err.Error(),
		}, nil
	} else if err != nil {
		return internalServerError(err), nil
	}
	slog.InfoContext(appCtx, "retrieved fi message object", "key", keyInput, "bucket", bucketInput)

	fiObjectData := make(map[string]jx.Raw)
	for k, v := range opsFiObject.Data {
		fiObjectData[k] = jx.Raw(v)
	}

	return &openapi.FiObject{
		ProfileId:   opsFiObject.ProfileId,
		Timestamp:   opsFiObject.Timestamp,
		OriginTopic: string(opsFiObject.OriginTopic),
		Data:        fiObjectData,
	}, nil
}